Split off an unburden-home-dir-doc package
This will make the unburden-home-dir package much smaller (.deb down
to 10kB from 85kB; installed size down to 28kB from 405kB) and is
recommended since Debian Policy version 3.9.7 anyways.

Additionally, fonts-font-awesome becomes a Recommends of
unburden-home-dir-doc instead of a Suggests of unburden-home-dir. With
the addition of a "Restrictions: needs-recommends" this will also fix
the adequate broken-symlink error if the Suggests wasn't installed.